backend/requirements.txt
fastapi
uvicorn[standard]
scikit-learn==1.6.1
numpy==2.0.2
pandas
joblib
shap
openai
prometheus-client
mlflow
redis
python-dotenv
tqdm
langchain>=0.0.200
google-genai
requests